rendertarget fixes

+++ Monosand/Graphics/SpriteBatch.cs
@@ -127,7 +127,7 @@ public sealed partial class SpriteBatch
         => DrawTexture(texture, position, Vector2.Zero, Vector2.One, 0f, Color.White);
 
     public void DrawRenderTarget(RenderTarget renderTarget, Vector2 position)
-        => DrawTexture(renderTarget.Texture, position + new Vector2(0f, (float)renderTarget.Texture.Height), new Vector2(1f, -1f));
+        => DrawTexture(renderTarget.Texture, position + new Vector2(0f, renderTarget.Texture.Height), new Vector2(1f, -1f));
 
     public void DrawTexture(Texture2D texture, Vector2 position, Color color)
         => DrawTexture(texture, position, Vector2.Zero, Vector2.One, 0f, color);
